requirements.txt
django==1.11
gunicorn==19.7.1
django-notifications-hq
djangorestframework==3.7.7
